Add cart component

+++ Breekly-FE/src/app/account/account.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Router, ActivatedRoute } from '@angular/router';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-account',
@@ -22,9 +23,26 @@ export class AccountComponent implements OnInit {
     )
   }
 
-  logout(){
-    this.router.navigateByUrl('')
-    localStorage.setItem("AUTH", "")
+  logout() {
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You will not be able to manage subscriptions anymore",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.router.navigateByUrl('')
+        localStorage.setItem("AUTH", "")
+        Swal.fire(
+          'Log out!',
+          'You were disconnected.',
+          'success'
+        )
+      }
+    })
   }
 
 }
